fix: revert optimize changes
Brings back old ES client usage, as migrating in
Optimize will be a bigger topic.

Reverts "refactor: migrate optimize client usage"

This reverts commit 90e9b915572cf9213075b8251aa6ba4638586d63.

+++ optimize/util/optimize-commons/src/main/java/io/camunda/optimize/service/db/es/schema/ElasticSearchIndexSettingsBuilder.java
@@ -48,10 +48,9 @@ public class ElasticSearchIndexSettingsBuilder {
                         m.nestedObjects(
                             n ->
                                 n.limit(
-                                    (long)
-                                        configurationService
-                                            .getElasticSearchConfiguration()
-                                            .getNestedDocumentsLimit()))));
+                                    configurationService
+                                        .getElasticSearchConfiguration()
+                                        .getNestedDocumentsLimit()))));
   }
 
   public static IndexSettings buildAllSettings(
@@ -77,10 +76,9 @@ public class ElasticSearchIndexSettingsBuilder {
                       m.nestedObjects(
                           n ->
                               n.limit(
-                                  (long)
-                                      configurationService
-                                          .getElasticSearchConfiguration()
-                                          .getNestedDocumentsLimit())));
+                                  configurationService
+                                      .getElasticSearchConfiguration()
+                                      .getNestedDocumentsLimit())));
           try {
             indexMappingCreator.getStaticSettings(i, configurationService);
           } catch (final IOException e) {
